requirements.txt
streamlit
pytesseract
Pillow
opencv-python-headless
numpy
pandas
openpyxl
